fix: 이벤트 신청 관련 수정(v4)
