+++ src/events/router/events.router.js
@@ -1,22 +1,123 @@
 // src/events/router/events.router.js
+/*
+  #swagger.components = {
+    securitySchemes: {
+      bearerAuth: { type: 'http', scheme: 'bearer', bearerFormat: 'JWT' }
+    }
+  }
+*/
 import { Router } from "express";
-
-// 하위 라우터들
-import creationRouter from "../creation/router/creation.router.js";
-import eventRouter from "../event/router/event.router.js";
-import applicationRouter from "../application/router/application.router.js";
-import restaurantsRouter from "../../restaurants/router/restaurants.router.js"; // /api/events/restaurants
+import { list, detail, edit, cancel } from "../service/event.service.js";
 
 const r = Router();
 
-// (선택) 디버그 로그
-console.log("[events] events.router.js loaded");
+/* 인증 미들웨어 (생략 가능: 이미 작성했던 버전 그대로) */
+let _authFn = null;
+async function authMw(req, res, next) {
+  try {
+    if (
+      process.env.NODE_ENV === "production" &&
+      process.env.SKIP_AUTH === "1"
+    ) {
+      const err = new Error("SKIP_AUTH must not be enabled in production");
+      err.status = 500;
+      return next(err);
+    }
+    if (
+      process.env.SKIP_AUTH === "1" &&
+      process.env.NODE_ENV !== "production"
+    ) {
+      if (!req.user) {
+        const fakeId = Number(process.env.DEV_FAKE_USER_ID || 1);
+        req.user = {
+          id: fakeId,
+          isCompleted: true,
+          nickname: "tester" + fakeId,
+        };
+        if (!global.__authBypassWarned) {
+          console.warn("[WARN] Auth bypass enabled (SKIP_AUTH=1) — dev only");
+          global.__authBypassWarned = true;
+        }
+      }
+      return next();
+    }
+    if (!_authFn) {
+      const mod = await import("../../../auth/middleware/auth.middleware.js");
+      const base = mod.authenticateAccessToken || mod.auth || mod.default;
+      if (typeof base !== "function") {
+        const err = new Error("AUTH_MIDDLEWARE_NOT_FOUND");
+        err.status = 500;
+        throw err;
+      }
+      _authFn = base;
+    }
+    return _authFn(req, res, (err) => {
+      if (err) return next(err);
+      if (!req.user && req.payload) {
+        const p = req.payload;
+        req.user = (p?.user || p) ?? null;
+        if (!req.user?.id) {
+          const e = new Error("UNAUTHORIZED");
+          e.status = 401;
+          return next(e);
+        }
+      }
+      next();
+    });
+  } catch (e) {
+    next(e);
+  }
+}
 
-// /api/events 하위에 매핑
-// ⚠️ 하위 라우터들은 전부 "루트('/') 기준"이어야 최종 경로가 /api/events/* 로 맞습니다.
-r.use("/", eventRouter); // 목록/상세/수정/삭제:  GET/PUT/DELETE /api/events[/:eventId]
-r.use("/", creationRouter); // 생성:               POST /api/events
-r.use("/", applicationRouter); // 신청/취소/내 신청:   /api/events/* (하위 라우터 정의에 따름)
-r.use("/restaurants", restaurantsRouter); // 식당 검색: /api/events/restaurants
+/* ──────────── 유틸 ──────────── */
+function parseEventId(req, res, next) {
+  const eventId = Number(req.params.eventId);
+  if (!Number.isInteger(eventId) || eventId <= 0) {
+    const err = new Error("INVALID_EVENT_ID");
+    err.status = 400;
+    return next(err);
+  }
+  req.eventId = eventId;
+  next();
+}
+
+/* ──────────── 라우터 ──────────── */
+r.get("/", async (req, res, next) => {
+  try {
+    const page = Number(req.query.page) || 1;
+    const size = Number(req.query.size) || 6;
+    const data = await list({ ...req.query, page, size });
+    return res.success ? res.success(data, 200) : res.status(200).json(data);
+  } catch (e) {
+    next(e);
+  }
+});
+
+r.get("/:eventId", parseEventId, async (req, res, next) => {
+  try {
+    const data = await detail(req.eventId);
+    return res.success ? res.success(data, 200) : res.status(200).json(data);
+  } catch (e) {
+    next(e);
+  }
+});
+
+r.put("/:eventId", authMw, parseEventId, async (req, res, next) => {
+  try {
+    const data = await edit(req.eventId, req.body, req.user);
+    return res.success ? res.success(data, 200) : res.status(200).json(data);
+  } catch (e) {
+    next(e);
+  }
+});
+
+r.delete("/:eventId", authMw, parseEventId, async (req, res, next) => {
+  try {
+    const data = await cancel(req.eventId, req.user);
+    return res.success ? res.success(data, 200) : res.status(200).json(data);
+  } catch (e) {
+    next(e);
+  }
+});
 
 export default r;
